moved all viewers to a data_viewer module and renamed files
