+++ src/pymodaq/daq_utils/calibration_camera.py
@@ -1,11 +1,11 @@
 from qtpy import QtWidgets
-from qtpy.QtCore import Qt, QObject, QLocale
+from qtpy.QtCore import Qt, QObject
 
 import pymodaq.daq_utils.parameter.ioxml
 from pyqtgraph.parametertree import Parameter, ParameterTree
 
 import sys
-from pymodaq.daq_utils.plotting.viewer2D.viewer2D_main import Viewer2D
+from pymodaq.daq_utils.plotting.data_viewers.viewer2D import Viewer2D
 from pymodaq.daq_measurement.daq_measurement_main import DAQ_Measurement
 from pymodaq.daq_utils.h5modules import browse_data
 import os
