+++ src/pymodaq/daq_utils/plotting/data_viewers/viewer0D_GUI.ui
