+++ src/pymodaq/daq_utils/plotting/viewer2D/__init__.py
@@ -1,1 +1,1 @@
-from .viewer2D_main import Viewer2D
+
