+++ src/pymodaq/daq_utils/plotting/viewer2D/viewer_2D_main.py
@@ -1,0 +1,4 @@
+from ...daq_utils import deprecation_msg
+from ..data_viewers.viewer2D import Viewer2D
+
+deprecation_msg('Viewer2D should now be imported from pymodaq.daq_utils.plotting.data_viewers')
