+++ src/pymodaq/resources/triangulation_data.npy
